refactor: Update StudentPreferences component to dynamically update team object
This commit updates the StudentPreferences component to dynamically update the team object based on the selected team ID. Previously, only the team ID was being updated, but now the team object includes additional properties such as the team name and survey. This change improves the functionality and accuracy of the component.